fix: Stop Stats monitor correctly when Track.onDispose.
