+++ lib/src/track/local/local.dart
@@ -205,7 +205,6 @@ abstract class LocalTrack extends Track {
     }
 
     logger.fine('$objectId.publish()');
-    startMonitor();
     _published = true;
     return true;
   }
@@ -219,26 +218,7 @@ abstract class LocalTrack extends Track {
     }
 
     logger.fine('$objectId.unpublish()');
-    stopMonitor();
     _published = false;
     return true;
-  }
-
-  Timer? _monitorTimer;
-
-  Future<void> monitorSender();
-
-  @internal
-  void startMonitor() {
-    _monitorTimer ??=
-        Timer.periodic(const Duration(milliseconds: monitorFrequency), (_) {
-      monitorSender();
-    });
-  }
-
-  @internal
-  void stopMonitor() {
-    _monitorTimer?.cancel();
-    _monitorTimer = null;
   }
 }
