+++ lib/src/track/remote/remote.dart
@@ -1,11 +1,9 @@
 import 'dart:async';
 
 import 'package:flutter_webrtc/flutter_webrtc.dart' as rtc;
-import 'package:meta/meta.dart';
 
 import '../../proto/livekit_models.pb.dart' as lk_models;
 import '../../types/other.dart';
-import '../stats.dart';
 import '../track.dart';
 
 abstract class RemoteTrack extends Track {
@@ -25,7 +23,6 @@ abstract class RemoteTrack extends Track {
     final didStart = await super.start();
     if (didStart) {
       await enable();
-      startMonitor();
     }
     return didStart;
   }
@@ -36,25 +33,6 @@ abstract class RemoteTrack extends Track {
     if (didStop) {
       await disable();
     }
-    stopMonitor();
     return didStop;
-  }
-
-  Timer? _monitorTimer;
-
-  Future<void> monitorReceiver();
-
-  @internal
-  void startMonitor() {
-    _monitorTimer ??=
-        Timer.periodic(const Duration(milliseconds: monitorFrequency), (_) {
-      monitorReceiver();
-    });
-  }
-
-  @internal
-  void stopMonitor() {
-    _monitorTimer?.cancel();
-    _monitorTimer = null;
   }
 }
