+++ lib/src/track/remote/video.dart
@@ -31,10 +31,10 @@ class RemoteVideoTrack extends RemoteTrack with VideoTrack {
   }
 
   @override
-  Future<void> monitorReceiver() async {
-    if (receiver == null) {
+  Future<bool> monitorStats() async {
+    if (receiver == null && events.isDisposed) {
       _currentBitrate = 0;
-      return;
+      return false;
     }
     final stats = await getReceiverStats();
 
@@ -45,6 +45,7 @@ class RemoteVideoTrack extends RemoteTrack with VideoTrack {
     }
 
     prevStats = stats;
+    return true;
   }
 
   Future<VideoReceiverStats?> getReceiverStats() async {
